moved draft related constants to handler package
