+++ cds-feature-attachments/src/main/java/com/sap/cds/feature/attachments/handler/draftservice/DraftConstants.java
@@ -1,8 +1,7 @@
 /**************************************************************************
  * (C) 2019-2024 SAP SE or an SAP affiliate company. All rights reserved. *
  **************************************************************************/
-package com.sap.cds.feature.attachments.handler.draftservice.constants;
-
+package com.sap.cds.feature.attachments.handler.draftservice;
 
 public final class DraftConstants {
