+++ cds-feature-attachments/src/main/java/com/sap/cds/feature/attachments/service/malware/DefaultAttachmentMalwareScanner.java
@@ -15,7 +15,7 @@ import org.slf4j.LoggerFactory;
 import com.sap.cds.CdsData;
 import com.sap.cds.Result;
 import com.sap.cds.feature.attachments.generated.cds4j.sap.attachments.Attachments;
-import com.sap.cds.feature.attachments.handler.draftservice.constants.DraftConstants;
+import com.sap.cds.feature.attachments.handler.draftservice.DraftConstants;
 import com.sap.cds.feature.attachments.service.AttachmentService;
 import com.sap.cds.feature.attachments.service.malware.client.MalwareScanClient;
 import com.sap.cds.feature.attachments.service.malware.client.mapper.MalwareClientStatusMapper;
@@ -27,11 +27,11 @@ import com.sap.cds.reflect.CdsEntity;
 import com.sap.cds.services.persistence.PersistenceService;
 
 /**
-	* The given input stream is scanned by using the malware scan service.
-	* The result of the scan	is returned with status {@link MalwareScanResultStatus}.
-	* In case of draft entity the data are tried to read from the draft entity first and if nothing is found
-	* the data are read from the main entity.
-	*/
+ * The given input stream is scanned by using the malware scan service.
+ * The result of the scan is returned with status {@link MalwareScanResultStatus}.
+ * In case of draft entity the data are tried to read from the draft entity first and if nothing is found
+ * the data are read from the main entity.
+ */
 public class DefaultAttachmentMalwareScanner implements AttachmentMalwareScanner {
 
 	private static final Logger logger = LoggerFactory.getLogger(DefaultAttachmentMalwareScanner.class);
